fix: consolidate migration formulas and add test (#1453)
* test: migration modals


* fix: consolidate migration formulas; close #1394
