+++ src/components/Migration/__tests__/MigrationModalSingle.unit.test.tsx
@@ -1,0 +1,9 @@
+import { render } from "../../../testUtils";
+import MigrationModalSingle from "../MigrationModalSingle";
+
+describe("<MigrationModalSingle/>", () => {
+  it("should render component", () => {
+    const { container } = render(<MigrationModalSingle open={false} handleClose={() => console.log("handleClose")} />);
+    expect(container).toMatchSnapshot();
+  });
+});
